+++ ajedrez/piece.py
@@ -1,7 +1,7 @@
 class Piece:
     def __init__(self, color):
         if color not in ["BLACK", "WHITE"]:
-            raise ValueError("El color debe ser 'BLACK' o 'WHITE'")
+            raise ValueError("El color debe ser 'BLACK' o 'WHITE'") #Por si no pone el color bien: crear excepción
         self.color = color
 
     def __str__(self):